EDK369
git-svn-id: https://svn.code.sf.net/p/efidevkit/code/trunk/Edk@154 d4c01ab0-1994-4ed3-9bc5-3ada14f9202f

+++ Sample/Universal/Variable/RuntimeDxe/Emu/EmuVariable.c
@@ -630,8 +630,8 @@ Returns:
     // NextVariable->NameSize should not include pad size so that variable
     // service can get actual size in GetVariable
     //
-    NextVariable->NameSize  = VarNameSize;
-    NextVariable->DataSize  = DataSize;
+    NextVariable->NameSize  = (UINT32)VarNameSize;
+    NextVariable->DataSize  = (UINT32)DataSize;
 
     EfiCopyMem (&NextVariable->VendorGuid, VendorGuid, sizeof (EFI_GUID));
     EfiCopyMem (
